MediAI-Api/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
motor==3.3.2
pymongo==4.5.0
pydantic==2.5.1
python-dotenv==1.0.1
pytest==7.4.2
httpx==0.25.1
